requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
python-dotenv
pymongo
certifi
pymongo[srv]==3.11
dill
pyaml
scipy
##-e .